Refactor StudentTable component and userController
This commit refactors the StudentTable component by removing the unnecessary useEffect hook and splitting the fetchData function into fetchStudentData and fetchCompaniesData. It also updates the fetchStudentData function to sort the students by rollNo and sets the students state accordingly. The fetchCompaniesData function fetches the companies and sets the companies state.

In the userController, the updateCompany function is updated to handle the case when the companyId is 'np' (Not Placed). It sets the placedAt object accordingly and updates the selectedStudentsRollNo array in the company if necessary.